finish/home

+++ controllers/articleController.js
@@ -9,14 +9,14 @@ module.exports.renderAddForm = function(req, res){
     };
     res.render('articles/add', {article});
 };
+
 module.exports.addArticle = async function(req, res){
-    console.log("Adding article now")
     const article = await Article.create( {
         title: req.body.title,
         intro: req.body.intro,
         image_url: req.body.image_url,
         body: req.body.body,
-        author_id: 1, //todo get logged in user
+        author_id: req.user.id,
         published_on: new Date()
     });
     res.redirect('/') //todo change the redirect to view all once made
@@ -52,10 +52,18 @@ module.exports.displayAll = async function(req, res){
 
 module.exports.renderEditForm = async function(req, res){
     const article = await Article.findByPk(req.params.articleId);
+    if (!article.isOwnedBy(user)){
+        res.redirect('/');
+        return;
+    }
     res.render('articles/edit', {article});
 };
 
 module.exports.updateArticle = async function(req, res){
+    const article = await Article.findByPk(req.params.articleId);
+    if (!article.isOwnedBy(user)){
+        res.redirect('/');
+    }
     await Article.update({
         title: req.body.title,
         intro: req.body.intro,
@@ -69,6 +77,11 @@ module.exports.updateArticle = async function(req, res){
     res.redirect(`/article/${req.params.articleId}`);
 }
 module.exports.deleteArticle = async function(req, res){
+    const article = await Article.findByPk(req.params.articleId);
+    if (!user.is('admin') && !article.isOwnedBy(user)){
+        res.redirect('/');
+        return;
+    }
     await Article.destroy({
         where: {
             id: req.params.articleId
